requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0 